--- src/main/java/nextstep/subway/applicaion/dto/LineUpdateRequest.java
@@ -1,9 +1,0 @@
-package nextstep.subway.applicaion.dto;
-
-import lombok.Getter;
-
-@Getter
-public class LineUpdateRequest {
-    private String name;
-    private String color;
-}
