--- src/main/java/nextstep/subway/applicaion/dto/StationRequest.java
@@ -1,11 +1,0 @@
-package nextstep.subway.applicaion.dto;
-
-import lombok.Getter;
-
-import javax.validation.constraints.NotBlank;
-
-@Getter
-public class StationRequest {
-    @NotBlank
-    private String name;
-}
