--- src/main/java/nextstep/subway/domain/LineRepository.java
@@ -1,10 +1,0 @@
-package nextstep.subway.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-
-public interface LineRepository extends JpaRepository<Line, Long> {
-    @Override
-    List<Line> findAll();
-}
