--- src/main/java/nextstep/subway/domain/Station.java
@@ -1,25 +1,0 @@
-package nextstep.subway.domain;
-
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-
-@Getter
-@Entity
-public class Station {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    private String name;
-
-    protected Station() {
-    }
-
-    public Station(String name) {
-        this.name = name;
-    }
-}
