--- src/main/java/nextstep/subway/domain/StationRepository.java
@@ -1,9 +1,0 @@
-package nextstep.subway.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-
-public interface StationRepository extends JpaRepository<Station, Long> {
-    List<Station> findByIdIn(List<Long> ids);
-}
