+++ src/main/java/nextstep/subway/line/application/dto/LineUpdateRequest.java
@@ -1,0 +1,9 @@
+package nextstep.subway.line.application.dto;
+
+import lombok.Getter;
+
+@Getter
+public class LineUpdateRequest {
+    private String name;
+    private String color;
+}
