+++ src/main/java/nextstep/subway/line/domain/LineRepository.java
@@ -1,0 +1,10 @@
+package nextstep.subway.line.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+
+public interface LineRepository extends JpaRepository<Line, Long> {
+    @Override
+    List<Line> findAll();
+}
