+++ src/main/java/nextstep/subway/station/applicaion/dto/StationRequest.java
@@ -1,0 +1,11 @@
+package nextstep.subway.station.applicaion.dto;
+
+import lombok.Getter;
+
+import javax.validation.constraints.NotBlank;
+
+@Getter
+public class StationRequest {
+    @NotBlank
+    private String name;
+}
